refactor(parser): split ParserState recovery helpers
